Error handling without try and catch
To get rid of uncaught errors we do not need try and catch block. Not if and error isn't thown. If we control de error ourselves we can also log it to the console with console.error. It will indicate that an error something went wrong. But because there is not an uncaught error, the program flow continue and the application does not crash.